pyproject.toml
[tool.poetry]
name = "vkv"
version = "0.1.0"
description = "Package to get and set keys on vault. Depends from vaultctl (https://github.com/fantunesdev/vaultctl)."
authors = ["Fernando Antunes <f.antunes84@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
hvac = "^1.0.2"
pynacl = "^1.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
